requirements.txt
# API and Environment
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5

# Data Processing
pymongo==4.5.0
pandas==2.0.3
numpy==1.24.3

# LLM and Evaluation
openai==1.4.0
langfuse==2.0.0
langchain==0.0.335
langchain-openai==0.0.2.post1

# Utilities
tqdm==4.66.1
pydantic==2.3.0 